refactor: remove sprite
